requirements.txt
langchain==0.3.25
langchain_community==0.3.23
langchain_core==0.3.58
langchain_groq==0.3.2
langchain_huggingface==0.1.2
python-dotenv==1.1.0
streamlit==1.44.0
